+++ qt/gen_qabstractitemmodel.go
@@ -778,7 +778,7 @@ func (this *QAbstractItemModel) Match4(start *QModelIndex, role int, value *QVar
 }
 
 func (this *QAbstractItemModel) Match5(start *QModelIndex, role int, value *QVariant, hits int, flags int) []QModelIndex {
-	var _ma *C.struct_miqt_array = C.QAbstractItemModel_Match5(this.h, start.cPointer(), (C.int)(role), value.cPointer(), (C.int)(hits), flags)
+	var _ma *C.struct_miqt_array = C.QAbstractItemModel_Match5(this.h, start.cPointer(), (C.int)(role), value.cPointer(), (C.int)(hits), (C.int)(flags))
 	_ret := make([]QModelIndex, int(_ma.len))
 	_outCast := (*[0xffff]*C.QModelIndex)(unsafe.Pointer(_ma.data)) // hey ya
 	for i := 0; i < int(_ma.len); i++ {
@@ -792,7 +792,7 @@ func (this *QAbstractItemModel) Match5(start *QModelIndex, role int, value *QVar
 }
 
 func (this *QAbstractItemModel) CheckIndex2(index *QModelIndex, options int) bool {
-	return (bool)(C.QAbstractItemModel_CheckIndex2(this.h, index.cPointer(), options))
+	return (bool)(C.QAbstractItemModel_CheckIndex2(this.h, index.cPointer(), (C.int)(options)))
 }
 
 func (this *QAbstractItemModel) DataChanged3(topLeft *QModelIndex, bottomRight *QModelIndex, roles []int) {
